+++ app-backend/db/src/main/scala/TaskDao.scala
@@ -2,6 +2,7 @@ package com.rasterfoundry.database
 
 import com.rasterfoundry.database.Implicits._
 import com.rasterfoundry.datamodel.GeoJsonCodec.PaginatedGeoJsonResponse
+import com.rasterfoundry.datamodel.Task.TaskPropertiesCreate
 import com.rasterfoundry.datamodel._
 
 import cats.data.OptionT
@@ -12,10 +13,12 @@ import doobie.postgres.implicits._
 import geotrellis.vector.{Geometry, Projected}
 import shapeless._
 
+import scala.concurrent.duration._
+
 import java.time.Instant
 import java.util.UUID
 
-object TaskDao extends Dao[Task] {
+object TaskDao extends Dao[Task] with ConnectionIOLogger {
 
   override val fieldNames = List(
     "id",
@@ -591,4 +594,48 @@ object TaskDao extends Dao[Task] {
            WHERE annotation_project_id = ${fromProject}
       """).update.run
   }
+
+  def getMostRecentStatus(task: Task): ConnectionIO[TaskStatus] =
+    getTaskActions(task.id) map { actions =>
+      actions.toNel map { actions =>
+        actions.tail.foldLeft(actions.head)(
+          (action1: TaskActionStamp, action2: TaskActionStamp) => {
+            if (action1.timestamp.isAfter(action2.timestamp)) action1
+            else action2
+          }
+        )
+      } map { _.fromStatus } getOrElse { TaskStatus.Unlabeled }
+    }
+
+  def expireStuckTasks(taskExpiration: FiniteDuration): ConnectionIO[Int] =
+    for {
+      _ <- info(s"Expiring stuck tasks: ${Instant.now}")
+      defaultUser <- UserDao.unsafeGetUserById("default")
+      stuckTasks <- query
+        .filter(
+          taskStatusF(
+            List(
+              TaskStatus.ValidationInProgress.repr,
+              TaskStatus.LabelingInProgress.repr
+            )
+          )
+        )
+        .filter(
+          fr"locked_on <= ${Instant.now.minusMillis(taskExpiration.toMillis)}"
+        )
+        .list
+      _ <- stuckTasks traverse { task =>
+        getMostRecentStatus(task) flatMap { status =>
+          val update =
+            Task.TaskFeatureCreate(
+              TaskPropertiesCreate(
+                status,
+                task.annotationProjectId
+              ),
+              task.geometry
+            )
+          updateTask(task.id, update, defaultUser)
+        }
+      }
+    } yield stuckTasks.length
 }